+++ delete_reply.php
@@ -3,6 +3,11 @@
     require_once("inc/header.php");
     require_once("app/models/Reply.php");
 
+    if(!$user->is_logged()) {
+        header("Location: login.php");
+        exit();
+    }
+
     if($_SERVER["REQUEST_METHOD"] == "POST") {
         $reply = new Reply();
         $reply = $reply->delete($_POST["reply_id"]);
